requirements.txt
Flask~=2.3.2
peewee~=3.16.2
psycopg2-binary~=2.9.6
python-dotenv~=1.0.0
gunicorn~=20.1.0
